requirements.txt
tensorflow==2.15.0
opencv-python==4.8.1.78
numpy==1.24.3
matplotlib==3.7.2
flask==2.3.3
pillow==10.0.1